refacto(Toggle)
