+++ UMI3D-BROWSER/Assets/Runtime/ContextualMenu/Private/02-Prefabs/ContextualMenu_Boolean.prefab
@@ -33,8 +33,8 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 6560147117377410631}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0, y: 1}
-  m_AnchorMax: {x: 0, y: 1}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 0, y: 0}
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
