+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Private/umi3d.browserRuntime.toggle.private.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "umi3d.browserRuntime.toggle.private",
+    "rootNamespace": "",
+    "references": [
+        "GUID:cade93c35bc88ed4d91a5492741b36ed"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
