+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Private/umi3d.browserRuntime.toggle.private.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 78b9cadfc0d21eb4991f1656e698d7ab
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
