+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Public.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2eb30de5d3f8fa64597578d216548e79
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
