+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Public/02-Prefabs/Toggle switch.prefab
@@ -33,8 +33,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.08, y: 0}
   m_AnchorMax: {x: 0.65, y: 1}
-  m_AnchoredPosition: {x: 0, y: 0.25}
-  m_SizeDelta: {x: -0.0000019073486, y: -1.0000019}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!1 &1867275513935416752
 GameObject:
@@ -70,8 +70,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
-  m_AnchoredPosition: {x: 22.866589, y: -0.0000076293945}
-  m_SizeDelta: {x: 54.1331, y: -81.3818}
+  m_AnchoredPosition: {x: 22.866585, y: 0}
+  m_SizeDelta: {x: 50, y: -80}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6187718521609448560
 CanvasRenderer:
@@ -94,14 +94,14 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Material: {fileID: 0}
-  m_Color: {r: 0.36862746, g: 0.89019614, b: 0.94117653, a: 1}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_RaycastTarget: 1
   m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
   m_Maskable: 1
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: 46f89b0c373bb274eb07f7c2e898fb40, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 9abc66eb44459d24da031b605f3706ce, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -148,8 +148,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 57.00354}
-  m_SizeDelta: {x: 160, y: 134.007}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 160, y: 130}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &465075924379518478
 MonoBehaviour:
@@ -198,7 +198,7 @@ MonoBehaviour:
   m_MinValue: 0
   m_MaxValue: 1
   m_WholeNumbers: 0
-  m_Value: 0
+  m_Value: 1
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
@@ -214,9 +214,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e7812998cbc30dd4a81319298a996e10, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  sliderValue: 0
-  animationDuration: 0
-  slideEase:
+  _isOn: 1
+  _animationDuration: 0.5
+  _slideEase:
     serializedVersion: 2
     m_Curve:
     - serializedVersion: 3
@@ -240,20 +240,12 @@ MonoBehaviour:
     m_PreInfinity: 2
     m_PostInfinity: 2
     m_RotationOrder: 4
-  onToggleOn:
-    m_PersistentCalls:
-      m_Calls: []
-  onToggleOff:
-    m_PersistentCalls:
-      m_Calls: []
-  backgroundImage: {fileID: 450023403433040610}
-  handleImage: {fileID: 588468839844156086}
-  recolorBackground: 1
-  recolorHandle: 1
-  backgroundColorOff: {r: 0.18431373, g: 0.18431373, b: 0.18431373, a: 0}
-  backgroundColorOn: {r: 0.36862746, g: 0.89019614, b: 0.94117653, a: 1}
-  handleColorOff: {r: 0.36862746, g: 0.89019614, b: 0.94117653, a: 1}
-  handleColorOn: {r: 0.18431373, g: 0.18431373, b: 0.18431373, a: 1}
+  _backgroundImage: {fileID: 450023403433040610}
+  _handleImage: {fileID: 588468839844156086}
+  _backgroundColorOff: {r: 1, g: 1, b: 1, a: 1}
+  _backgroundColorOn: {r: 0, g: 0, b: 0, a: 1}
+  _handleColorOff: {r: 0, g: 0, b: 0, a: 1}
+  _handleColorOn: {r: 1, g: 1, b: 1, a: 1}
 --- !u!65 &4246181915148262145
 BoxCollider:
   m_ObjectHideFlags: 0
@@ -333,7 +325,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Material: {fileID: 0}
-  m_Color: {r: 0.18431373, g: 0.18431373, b: 0.18431373, a: 0}
+  m_Color: {r: 0, g: 0, b: 0, a: 1}
   m_RaycastTarget: 1
   m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
   m_Maskable: 1
