+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Public/04-Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd32bc86865c74d4785a3b217f99700e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
