+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Public/umi3d.browserRuntime.toggle.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "umi3d.browserRuntime.toggle"
+}
