+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Public/umi3d.browserRuntime.toggle.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cade93c35bc88ed4d91a5492741b36ed
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
