+++ UMI3D-BROWSER/Assets/Runtime/Toggle/Tests/umi3d.browserRuntime.toggle.tests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4d470d9c9f97ee2448069d1be1ffd931
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
